# Speed up `BinningVectorizer`