First, thanks for the great library!

I was playing with methods for the DRIAMS-A part of the data and noticed that `BinningVectorizer` was using a lot of time. Specifically, I load all years, species and Ciprofloxacin + Penicillin:
```python
explorer = DRIAMSDatasetExplorer(DRIAMS_ROOT)

driams_dataset = load_driams_dataset(
            explorer.root,
            'DRIAMS-A',
            ['2015', '2016', '2017', '2018'],
            '*',
            ['Ciprofloxacin', 'Penicillin'],
            handle_missing_resistance_measurements='remove_if_all_missing',
            id_suffix='strat',
)
len(driams_dataset.X)  # 33415
```

Using the `BinningVectorizer` with 200 bins then takes 9 seconds for fitting and over 6 minutes for the transform:
```python
bv = BinningVectorizer(200, min_bin=2000, max_bin=20000)
bv.fit(driams_dataset.X)               # 8.9s
Xold = bv.transform(driams_dataset.X)  # 6m 10.0s
```
The time also seems to worsen as I increase the bin count.

I tried to optimize the code for `fit` and `transform` and I can now get results significantly faster:
```python
nbv = NewBinningVectorizer(200, min_bin=2000, max_bin=20000)
nbv.fit(driams_dataset.X)               # 2.1s
Xnew = nbv.transform(driams_dataset.X)  # 20.8s
```
The results are also the same (up to small rounding errors):
```python
np.all(mbv.bin_edges_ == bv.bin_edges_)         # True
np.abs(Xnew - Xold).max()                       # abs. error = 4.104762319975808e-15
np.abs(Xnew - Xold).max() / np.abs(Xold).max()  # rel. error = 1.3664372704227028e-14
```
It also passes the test in `tests/vectorization/test_binning_vectorizer.py`.

I haven't thoroughly checked the performance for smaller datasets. However, with the smaller example in the README (1520 samples) the new code is still faster.

Hopefully, this is useful :) If anything looks weird or should be changed please let me know! I tried to explain each change in their respective commits.